Improve use of states
Even if states can be handled by Thelia, many functions which should use them do not get them as parameter,
in particular filtering addresses on delivering. This commit fixes those calls.
Question about the findByCountryAndState() function of the CountryAreaQuery class: if there is a state given,
should areas with no state also be returned?
Question about the foreach in the findByCountryAndModule() function of the AreaDeliveryModuleQuery class:
is the findOne() really appropriate? Does it return the good area?